+++ ts/packages/spl-name-service/rollup.config.ts
@@ -38,7 +38,6 @@ export default {
     "bn.js",
     "bs58",
     "buffer",
-    "camelcase",
     "eventemitter3",
     "@noble/hashes/sha256",
     "pako",